Bug in change of scenario fixed
A bug in the change of the scenario script has been fixed. By pressing twice the space bar when over a change-scenario arrow the script was enqueued twice and the character was moved to the default position of the next scenario. The bug prevents the use to press more than once the space button.